refactor(core): Extract community package service from controller
Extract shared install/update/uninstall logic into a dedicated lifecycle
service so it can be reused by public API
